refactor: 헤더 컴포넌트 분리 및 컴포넌트 폴더 구조 수정
+스티키 헤더 효과 제거
