+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Header from '@/components/Header/Header';
-import Footer from '@/components/Footer/Footer';
+import Header from '@/components/common/Header/Header';
+import Footer from '@/components/common/Footer/Footer';
 import { ScrollRestoration } from 'react-router-dom';
 
 function App() {
