+++ src/components/common/ImgWrapper/ImgWrapper.stories.ts
