+++ src/pages/ErrorPage/ErrorPage.tsx
@@ -1,7 +1,7 @@
 import * as S from './ErrorPage.styles';
 import { useRouteError, isRouteErrorResponse } from 'react-router-dom';
-import Header from '@/components/Header/Header';
-import Footer from '@/components/Footer/Footer';
+import Header from '@/components/common/Header/Header';
+import Footer from '@/components/common/Footer/Footer';
 
 function ErrorPage() {
   const error = useRouteError();
