+++ src/types/NavigationType.ts
@@ -1,0 +1,4 @@
+export interface Route {
+  path: string;
+  label: string;
+}
