+++ src/utils/navigationUtils.ts
@@ -1,0 +1,5 @@
+import { Route } from '@/types/NavigationType';
+
+export const getFilteredRoutes = (routes: Route[], targetIdx: number) => {
+  return routes.slice(targetIdx);
+};
